mySqlConnect/requirements.txt
mypy==0.910
mypy-extensions==0.4.3
PyMySQL==1.0.2
python-dotenv==0.19.0
toml==0.10.2
typing-extensions==3.10.0.2
